internal/linux: Move UTS/Uname functions
